add init script when overclocking that allows the user to opt out on boot
Current kernel builds default to the powersave cpufreq governor, with the
intention that this is changed to 'ondemand' after boot. When the user sets
overclock in raspi-config, an init script is installed which does this switch,
but only does it if it doesn't detect a shift key is held down. This means the
user can still boot the system if they set an overclock setting that is higher
than their device can manage.